task4/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
pandas
python-dotenv
playwright